cdk/lambda/requirements.txt
PyJWT[crypto]==2.8.0
boto3==1.34.89
